SpiralWayGenerator

+++ src/rocnikovkapc/WayGenerator.java
@@ -21,7 +21,7 @@ public class WayGenerator {
     }
 
     public Waypoint gnw(Waypoint actual) {
-        boolean nahoru = Math.round(actual.x / (TRACK_WIDTH / 2)) == 1 || Math.round(actual.x / (TRACK_WIDTH / 2)) % 5 == 0 ? true : false;
+        boolean nahoru = Math.round(((double)actual.x) / (TRACK_WIDTH / 2.0)) == 1 || Math.round(actual.x / (TRACK_WIDTH / 2)) % 5 == 0 ? true : false;
         if (nahoru) {
             if (Math.abs(YRANGE - actual.y) < TRACK_WIDTH * 1.25) {
                 return new Waypoint(Math.min(XRANGE - TRACK_WIDTH, actual.x + STEP_LENGTH), actual.y);
